refactor: 매직넘버 상수화, util 로직 분리
