+++ src/components/Card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CardNumber } from '../type';
-import { changeNumberToMask } from '../utils/util';
+import { changeNumberToMask } from '../utils/processData';
 import './Card.css';
 
 type CardProps = {
