+++ src/utils/util.ts
@@ -1,53 +1,3 @@
-const splitStringIntoGroups = (str: string, groupSize: number): string[] => {
-  const result = [];
-  let temp = [];
-
-  for (let index = 0; index < str.length; index++) {
-    if (index !== 0 && (index + 1) % groupSize === 0) {
-      temp.push(str[index]);
-
-      result.push(temp.join(''));
-      temp = [];
-    }
-    temp.push(str[index]);
-  }
-  return result;
-};
-
-export const formatCardNumber = (cardNumber: string): string => {
-  if (cardNumber.length <= 4) return cardNumber;
-  const splitCardNumber = splitStringIntoGroups(cardNumber, 4)
-    .map((number, index) => {
-      if (index > 1) return '*'.repeat(number.length);
-      return number;
-    })
-    .join('-');
-  return splitCardNumber;
-};
-
-const MONTH_DATA = ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'];
-export const formatExpireDate = (expireDate: string): string => {
-  if (expireDate[0] !== '0' && expireDate[0] !== '1') return '';
-  if (expireDate.length > 1 && !MONTH_DATA.includes(`${expireDate[0]}${expireDate[1]}`)) return '';
-
-  const nowLength = expireDate.length;
-  const nowString = expireDate.split('');
-  if (nowLength === 3 && nowString.includes('/')) {
-    return expireDate.slice(0, -1);
-  }
-  if (nowLength === 3) {
-    return `${expireDate[0]}${expireDate[1]}/${expireDate[2]}`;
-  }
-  return expireDate;
-};
-
-export const handleNumberInput = (data: string): string => {
-  if (!/[0-9]/.test(data[data.length - 1])) {
-    data = data.slice(0, -1);
-  }
-  return data;
-};
-
 export const isAlphabetInput = (data: string): boolean => {
   return /[A-Z]/g.test(data);
 };
@@ -56,31 +6,6 @@ export const isNumberInput = (data: string): boolean => {
   return /[0-9]/g.test(data);
 };
 
-export const changeNumberToMask = (data: string): string => {
-  return '·'.repeat(data.length);
-};
-
-export const cardExpireCondition = (e: React.ChangeEvent<HTMLInputElement>) => {
-  const length = e.target.value.length;
-  const lastWord = length === 0 ? '' : e.target.value[length - 1];
-  return (
-    length <= 5 && length >= 0 && (isNumberInput(lastWord) || lastWord === '/' || lastWord === '')
-  );
-};
-
-export const securityCodeCondition = (e: React.ChangeEvent<HTMLInputElement>) => {
-  return e.target.value.length <= 3;
-};
-export const cardOwnerCondition = (e: React.ChangeEvent<HTMLInputElement>) => {
-  const length = e.target.value.length;
-  const lastWord = e.target.value[length - 1];
-  return length <= 30 && !(length > 0 && !isAlphabetInput(lastWord.toUpperCase()));
-};
-
 export const stringToUpperCase = (data: string): string => {
   return data.toUpperCase();
-};
-
-export const cardPasswordCondition = (e: React.ChangeEvent<HTMLInputElement>) => {
-  return e.target.value.length <= 1;
 };
